Cleanup props, refactor formatLabel
